Feature 0.0.1/zsm test

See merge request hs-iot/iot-platform!148

+++ dao/src/main/java/org/thingsboard/server/dao/hs/service/DeviceMonitorServiceImpl.java
@@ -5,6 +5,7 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.curator.shaded.com.google.common.collect.Sets;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
@@ -128,7 +129,7 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
     public DeviceProfileVO getDeviceProfileDetail(TenantId tenantId, DeviceProfileId deviceProfileId) {
         var deviceProfile = deviceProfileService.findDeviceProfileById(tenantId, deviceProfileId);
 
-        var dictDeviceList = DaoUtil.convertDataList(this.deviceProfileDictDeviceRepository.findAllBindDeviceProfile(deviceProfile.getId().getId()));
+        var dictDeviceList = DaoUtil.convertDataList(this.deviceProfileDictDeviceRepository.findAllDictDeviceEntityByDeviceProfileId(deviceProfile.getId().getId()));
 
         DeviceProfileVO deviceProfileVO = new DeviceProfileVO();
         BeanUtils.copyProperties(deviceProfile, deviceProfileVO);
@@ -159,24 +160,24 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
     @Transactional
     public void updateAlarmStatus(TenantId tenantId, AlarmId alarmId, long ts, AlarmStatus alarmStatus) throws ThingsboardException {
         Alarm alarm = this.alarmRepository.findByTenantIdAndId(tenantId.getId(), alarmId.getId())
-                .map(AlarmEntity::toData).orElseThrow(() -> new ThingsboardException("TenantId error！", ThingsboardErrorCode.GENERAL));
+                .map(AlarmEntity::toData).orElseThrow(() -> new ThingsboardException("该报警信息不存在！", ThingsboardErrorCode.GENERAL));
         switch (alarmStatus) {
             case ACTIVE_ACK:
                 if (!AlarmSimpleStatus.valueOf(alarm.getStatus().toString()).canBeConfirm()) {
-                    throw new ThingsboardException("alarm status is not ACTIVE_UNACK！", ThingsboardErrorCode.GENERAL);
+                    throw new ThingsboardException("该报警状态非未确认！", ThingsboardErrorCode.GENERAL);
                 }
                 alarm.setAckTs(ts);
                 break;
             case CLEARED_ACK:
                 if (!AlarmSimpleStatus.valueOf(alarm.getStatus().toString()).canBeClear()) {
-                    throw new ThingsboardException("alarm status is not ACTIVE_UNACK or ACTIVE_ACK！", ThingsboardErrorCode.GENERAL);
+                    throw new ThingsboardException("该报警状态非未确认或已确认！", ThingsboardErrorCode.GENERAL);
                 }
                 if (alarm.getAckTs() == 0L)
                     alarm.setAckTs(ts);
                 alarm.setClearTs(ts);
                 break;
             default:
-                throw new ThingsboardException("alarm status is not support", ThingsboardErrorCode.GENERAL);
+                throw new ThingsboardException("该报警状态不支持！", ThingsboardErrorCode.GENERAL);
         }
         alarm.setStatus(alarmStatus);
         this.alarmRepository.save(new AlarmEntity(alarm));
@@ -267,6 +268,7 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                 .offLineDeviceCount(allDeviceList.size() - onLineCount)
                 .onLineDeviceCount(onLineCount)
                 .alarmTimesList(this.listAlarmTimesResult(tenantId, allDeviceIdList))
+                .deviceIdList(allDeviceIdList.stream().map(UUID::toString).collect(Collectors.toList()))
                 .build();
     }
 
@@ -279,27 +281,25 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
      */
     @Override
     public DeviceDetailResult getRTMonitorDeviceDetail(TenantId tenantId, String id) throws ExecutionException, InterruptedException, ThingsboardException {
-        var device = Optional.ofNullable(this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id))).map(DeviceEntity::toData).orElseThrow(() -> new ThingsboardException("device not exist", ThingsboardErrorCode.GENERAL));
+        var device = Optional.ofNullable(this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id))).map(DeviceEntity::toData).orElseThrow(() -> new ThingsboardException("设备不存在", ThingsboardErrorCode.GENERAL));
         var deviceBaseDTO = this.clientService.getDeviceBase(tenantId, new FactoryDeviceQuery(UUIDToString(device.getFactoryId()), UUIDToString(device.getWorkshopId()), UUIDToString(device.getProductionLineId()), device.getId().toString()));
 
         var kvEntryMap = this.timeseriesService.findAllLatest(tenantId, DeviceId.fromString(id)).get()
                 .stream().sorted(Comparator.comparing(TsKvEntry::getKey)).collect(Collectors.toMap(TsKvEntry::getKey, Function.identity(), (key1, key2) -> key1, LinkedHashMap::new));
 
-        var dictDataMap = this.dictDataService.listDictDataByKeys(tenantId,
-                kvEntryMap.entrySet().stream().reduce(new ArrayList<>(), (r, e) -> {
-                    r.add(e.getKey());
-                    return r;
-                }, (a, b) -> null));
+        Map<String, DictData> dictDataMap = this.dictDataService.mapAllDictData(tenantId);
 
         List<DictDeviceGroupVO> groupResultList = new ArrayList<>();
+        List<DictDeviceComponentVO> componentList = new ArrayList<>();
         List<String> groupPropertyNameList = new ArrayList<>();
 
         DictDevice dictDevice = new DictDevice();
         if (device.getDictDeviceId() != null) {
             dictDevice = this.dictDeviceRepository.findByTenantIdAndId(tenantId.getId(), device.getDictDeviceId()).map(DictDeviceEntity::toData).orElseGet(DictDevice::new);
 
-            var groupVOList = this.dictDeviceService.listDictDeviceGroup(device.getDictDeviceId());
+            var dictDeviceDetail = this.dictDeviceService.getDictDeviceDetail(device.getDictDeviceId().toString(), tenantId);
 
+            var groupVOList = dictDeviceDetail.getGroupList();
             groupResultList = groupVOList.stream().map(e -> DictDeviceGroupVO.builder()
                     .id(e.getId())
                     .name(e.getName())
@@ -308,7 +308,7 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                         kvData.ifPresent(k -> groupPropertyNameList.add(v.getName()));
                         return DictDeviceGroupPropertyVO.builder()
                                 .id(v.getId())
-                                .unit(Optional.ofNullable(dictDataMap.get(v.getName())).map(DictData::getUnit).orElse(null))
+                                .unit(Optional.ofNullable(v.getDictDataId()).map(dictDataMap::get).map(DictData::getUnit).orElse(null))
                                 .name(v.getName())
                                 .title(v.getTitle())
                                 .content(kvData.isEmpty() ? v.getContent() : kvData.get().getValue().toString())
@@ -316,6 +316,9 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                                 .build();
                     }).collect(Collectors.toList()))
                     .build()).collect(Collectors.toList());
+
+            componentList = dictDeviceDetail.getComponentList();
+            this.recursionDealComponentData(componentList, kvEntryMap, dictDataMap, groupPropertyNameList);
         }
 
         var ungrouped = DictDeviceGroupVO.builder().name(HSConstants.UNGROUPED).groupPropertyList(new ArrayList<>()).build();
@@ -330,8 +333,6 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                         .build());
             }
         });
-        if (!ungrouped.getGroupPropertyList().isEmpty())
-            groupResultList.add(ungrouped);
 
         return DeviceDetailResult.builder()
                 .id(device.getId().toString())
@@ -343,6 +344,8 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                 .productionLineName(Optional.ofNullable(deviceBaseDTO.getProductionLine()).map(ProductionLine::getName).orElse(null))
                 .isUnAllocation(this.isDeviceUnAllocation(device))
                 .resultList(groupResultList)
+                .componentList(componentList)
+                .resultUngrouped(ungrouped)
                 .alarmTimesList(this.listAlarmTimesResult(tenantId, List.of(toUUID(id))))
                 .build();
     }
@@ -358,7 +361,8 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
      * @return 设备分组属性历史数据
      */
     @Override
-    public List<DictDeviceGroupPropertyVO> listGroupPropertyHistory(TenantId tenantId, String deviceId, String groupPropertyName, Long startTime, Long endTime) throws ExecutionException, InterruptedException {
+    @SuppressWarnings("Duplicates")
+    public HistoryVO listGroupPropertyHistory(TenantId tenantId, String deviceId, String groupPropertyName, Long startTime, Long endTime) throws ExecutionException, InterruptedException {
         List<String> keyList = new ArrayList<>() {{
             add(groupPropertyName);
         }};
@@ -368,36 +372,29 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
         var tempResult = this.timeseriesService.findAll(tenantId, DeviceId.fromString(deviceId), tempQueries).get()
                 .stream().collect(Collectors.toMap(TsKvEntry::getKey, Function.identity()));
         if (tempResult.isEmpty())
-            return Lists.newArrayList();
+            return HistoryVO.builder().isShowChart(false).propertyVOList(Lists.newArrayList()).build();
         int count = Integer.parseInt(String.valueOf(tempResult.get(groupPropertyName).getValue()));
         List<ReadTsKvQuery> queries = keyList.stream().map(key -> new BaseReadTsKvQuery(key, startTime, endTime, endTime - startTime, count, Aggregation.NONE, "desc"))
                 .collect(Collectors.toList());
 
-        var dictDataMap = this.dictDataService.listDictDataByKeys(tenantId, keyList);
+        var dictDataMap = this.dictDataService.mapAllDictData(tenantId);
+        Map<String, String> rMap = Maps.newHashMap();
+        var device = Optional.ofNullable(this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(deviceId))).map(DeviceEntity::toData).orElse(null);
+        if (device != null && device.getDictDeviceId() != null) {
+            rMap = this.dictDeviceService.mapAllPropertyDictDataId(device.getDictDeviceId());
+        }
+        Map<String, String> finalRMap = rMap;
 
-        return this.timeseriesService.findAll(tenantId, DeviceId.fromString(deviceId), queries).get()
+        var data = this.timeseriesService.findAll(tenantId, DeviceId.fromString(deviceId), queries).get()
                 .stream().map(e -> DictDeviceGroupPropertyVO.builder()
                         .content(e.getValue().toString())
-                        .unit(Optional.ofNullable(dictDataMap.get(groupPropertyName)).map(DictData::getUnit).orElse(null))
+                        .unit(Optional.ofNullable(finalRMap.getOrDefault(groupPropertyName, null))
+                                .map(dictDataMap::get).map(DictData::getUnit).orElse(null))
                         .createdTime(e.getTs())
-                        .build()).collect(Collectors.toList());
-    }
+                        .build())
+                .collect(Collectors.toList());
 
-    /**
-     * 【APP】查询设备分组属性历史数据
-     *
-     * @param tenantId          租户Id
-     * @param deviceId          设备Id
-     * @param groupPropertyName 属性名称
-     * @param startTime         开始时间
-     * @param endTime           结束时间
-     * @return 设备分组属性历史数据
-     */
-    @Override
-    public AppHistoryVO listAppGroupPropertyHistory(TenantId tenantId, String deviceId, String groupPropertyName, Long startTime, Long endTime) throws ExecutionException, InterruptedException {
-        var data = this.listGroupPropertyHistory(tenantId, deviceId, groupPropertyName, startTime, endTime);
-
-        return AppHistoryVO.builder()
+        return HistoryVO.builder()
                 .propertyVOList(data)
                 .isShowChart(!data.isEmpty() && isNumberData(data.get(0).getContent()) ? Boolean.TRUE : Boolean.FALSE)
                 .build();
@@ -421,7 +418,7 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
         var temp = this.timeseriesService.findAll(tenantId, DeviceId.fromString(deviceId), tempQueries).get()
                 .stream().map(KvEntry::getValue).map(e -> Integer.valueOf(String.valueOf(e))).collect(Collectors.toList());
         if (temp.isEmpty())
-            new PageData<>(Lists.newArrayList(), 0, 0L, false);
+            return new PageData<>(Lists.newArrayList(), 0, 0L, false);
         var count = temp.stream().mapToInt(Integer::intValue).sum();
         var max = temp.stream().max(Integer::compareTo).orElse(0);
 
@@ -450,6 +447,7 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
      * @return 查询设备历史-表头，包含时间
      */
     @Override
+    @SuppressWarnings("Duplicates")
     public List<DictDeviceGroupPropertyVO> listDictDeviceGroupPropertyTitle(TenantId tenantId, String deviceId) {
         List<DictDeviceGroupPropertyVO> propertyVOList = new ArrayList<>() {{
             add(DictDeviceGroupPropertyVO.builder()
@@ -459,12 +457,20 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
         var keyList = this.timeseriesService.findAllKeysByEntityIds(tenantId, List.of(DeviceId.fromString(deviceId)));
         if (keyList.isEmpty())
             return propertyVOList;
-        var dictDataMap = this.dictDataService.listDictDataByKeys(tenantId, keyList);
+
+        var dictDataMap = this.dictDataService.mapAllDictData(tenantId);
+        Map<String, String> rMap = Maps.newHashMap();
+        var device = Optional.ofNullable(this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(deviceId))).map(DeviceEntity::toData).orElse(null);
+        if (device != null && device.getDictDeviceId() != null) {
+            rMap = this.dictDeviceService.mapAllPropertyDictDataId(device.getDictDeviceId());
+        }
+        Map<String, String> finalRMap = rMap;
 
         propertyVOList.addAll(keyList.stream().map(e -> DictDeviceGroupPropertyVO.builder()
                 .name(e)
                 .title(e)
-                .unit(Optional.ofNullable(dictDataMap.get(e)).map(DictData::getUnit).orElse(null))
+                .unit(Optional.ofNullable(finalRMap.getOrDefault(e, null))
+                        .map(dictDataMap::get).map(DictData::getUnit).orElse(null))
                 .build()).collect(Collectors.toList()));
         return propertyVOList;
     }
@@ -532,14 +538,33 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
     /**
      * 绑定设备字典到设备配置
      *
+     * @param tenantId         租户Id
      * @param dictDeviceIdList 设备字典Id列表
      * @param deviceProfileId  设备配置Id
      */
     @Override
     @Transactional
-    public void bindDictDeviceToDeviceProfile(List<String> dictDeviceIdList, DeviceProfileId deviceProfileId) {
-        this.deleteBindDictDevice(deviceProfileId);
-        deviceProfileDictDeviceRepository.saveAll(dictDeviceIdList.stream().map(e -> {
+    public void bindDictDeviceToDeviceProfile(TenantId tenantId, List<String> dictDeviceIdList, DeviceProfileId deviceProfileId) {
+        // 在其它程序代码无错误的前提下，设备正确绑定了设备配置的情况下执行下列代码
+        var deleteList = Sets.newHashSet(deviceProfileDictDeviceRepository.findAllByDeviceProfileId(deviceProfileId.getId()).stream()
+                .map(DeviceProfileDictDeviceEntity::getDictDeviceId).collect(Collectors.toList()));
+        var addList = Sets.newHashSet(dictDeviceIdList.stream().map(this::toUUID).collect(Collectors.toList()));
+        if (addList.equals(deleteList))
+            return;
+        var defaultDeviceProfileId = Optional.ofNullable(this.deviceProfileRepository.findByDefaultTrueAndTenantId(tenantId.getId())).map(DeviceProfileEntity::getId).orElse(null);
+
+        var deleteDeviceList = this.deviceRepository.findAllByTenantIdAndDictDeviceIdIn(tenantId.getId(),
+                deleteList.stream().filter(e -> !addList.contains(e)).collect(Collectors.toSet()));
+        deleteDeviceList.forEach(e -> e.setDeviceProfileId(defaultDeviceProfileId));
+        this.deviceRepository.saveAll(deleteDeviceList);
+
+        var addDeviceList = this.deviceRepository.findAllByTenantIdAndDictDeviceIdIn(tenantId.getId(),
+                addList.stream().filter(e -> !deleteList.contains(e)).collect(Collectors.toSet()));
+        addDeviceList.forEach(e -> e.setDeviceProfileId(deviceProfileId.getId()));
+        this.deviceRepository.saveAll(addDeviceList);
+
+        this.deleteBindDictDevice(deviceProfileId); // default
+        deviceProfileDictDeviceRepository.saveAll(dictDeviceIdList.stream().map(e -> {  // default
             DeviceProfileDictDeviceEntity entity = new DeviceProfileDictDeviceEntity();
             entity.setDictDeviceId(toUUID(e));
             entity.setDeviceProfileId(deviceProfileId.getId());
@@ -570,6 +595,32 @@ public class DeviceMonitorServiceImpl extends AbstractEntityService implements D
                     .time(month)
                     .num(Optional.ofNullable(alarmMap.get(e)).map(List::size).orElse(0)).build();
         }).collect(Collectors.toList());
+    }
+
+    /**
+     * 递归处理部件属性数据
+     *
+     * @param componentList         部件列表
+     * @param kvEntryMap            遥测数据map
+     * @param dictDataMap           数据字典map
+     * @param groupPropertyNameList 属性List
+     */
+    public void recursionDealComponentData(List<DictDeviceComponentVO> componentList, Map<String, TsKvEntry> kvEntryMap, Map<String, DictData> dictDataMap, List<String> groupPropertyNameList) {
+        for (DictDeviceComponentVO componentVO : componentList) {
+            for (DictDeviceComponentPropertyVO propertyVO : componentVO.getPropertyList()) {
+                var kvData = Optional.ofNullable(kvEntryMap.get(propertyVO.getName()));
+                kvData.ifPresent(k -> {
+                    groupPropertyNameList.add(propertyVO.getName());
+                    propertyVO.setContent(kvData.get().getValue().toString());
+                    propertyVO.setUnit(Optional.ofNullable(propertyVO.getDictDataId()).map(dictDataMap::get).map(DictData::getUnit).orElse(null));
+                    propertyVO.setCreatedTime(kvData.get().getTs());
+                });
+            }
+            if (componentVO.getComponentList() == null || componentVO.getComponentList().isEmpty()) {
+                continue;
+            }
+            this.recursionDealComponentData(componentVO.getComponentList(), kvEntryMap, dictDataMap, groupPropertyNameList);
+        }
     }
 
     @Autowired
